+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintExtension.kt
@@ -56,6 +56,10 @@ internal constructor(
      */
     val coloredOutput: Property<Boolean> = objectFactory.property { set(true) }
     /**
+     * Specify the color of the terminal output.
+     */
+    val outputColorName: Property<String> = objectFactory.property { set("") }
+    /**
      * Whether or not to allow the build to continue if there are warnings;
      * defaults to {@code false}, as for any other static code analysis tool.
      * <p>